refactor: re-use some code from filterWeb3Tx to parse-tx
